refactor(web): rename SettingsContent to AppearanceTab
Move settings content into settings/ directory and rename from
SettingsContent to AppearanceTab to align with upcoming tab-based
settings structure. Update import in SettingsMenu accordingly.
